requirements/requirements.txt
flask==2.1.2
flask_login==0.5.0
flask_migrate==2.7.0
flask_wtf==0.14.3
flask_sqlalchemy==2.4.4
sqlalchemy==1.3.23
email_validator==1.1.2
python-decouple==3.4
gunicorn==20.1.0
numpy==1.24.2
cython==0.29.33
pandas==1.5.3
pathlib==1.0.1
python-dotenv==0.21.1
matplotlib==3.7.0
#flask-restful
flask-apispec==0.11.4
flask-restx==1.0.6
PyJWT==2.6.0
Werkzeug~=2.0.0
Flask-Session==0.4.0
Flask-Admin==1.5.8
Jinja2==3.0.3
## for dump
flask-alchemydumps==0.0.13
sqlite-dump==0.1.1
